requirements.txt
langchain
google-generativeai
duckduckgo-search
fpdf
python-dotenv
